requirements.txt
streamlit
python-dotenv
langchain_community
langchain
langchain-text-splitters
langchain_pinecone
pinecone[grpc]
pinecone-client
assemblyai
langchain-groq
nltk
google.api
google-api-core 
protobuf
protoc-gen-openapiv2
grpcio
google-generativeai
sentence-transformers